+++ utils/functions.go
@@ -2,7 +2,7 @@ package utils
 
 /**
  * The [StringInStrings] function determines whether a particular string is
- * contained within an array of strings.
+ * contained within a slice of strings.
  */
 func StringInStrings (search string, stringsArr []string) bool {
   for _, stringItem := range stringsArr {